[DOCS] Properly reflect Fluid "value" resolving in inline comment
If specifying both a "property" and a "value" attribute to a
f:form based Fluid ViewHelper like f:form.textfield, there
is potentially a conflict on which values to use.

It has been discussed that providing a "value" should
have preference over a property value, which is the
current behaviour. However, the inline code comment
did not reflect that.

It is now adjusted to match the actual implementation.

Resolves: #99243
Releases: main
Change-Id: Ib6bec70302e461f3e8d70ac41605a1d24e3d7a25
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85413
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>