[BUGFIX] Allow non-english languages as xlf default language
By allowing `en` as language key and not falling back to `default`
automatically, integrators can get back an edge case to use
a different language in default XLF files and provide an english
translation in `en.locallang.xlf` files by registering English as
another language.

This is not the suggested way but has worked in 11.5 and is a valid
use case if clients use a different language as basis for e.g.
translation agencies.

Resolves: #104468
Releases: main, 12.4
Change-Id: Id35c7fd1e3a6b50d2b19eaa418fa4a2ba8c46024
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85881
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>