[TASK] Dynamically limit livesearch to searchFields used in subtypes
Previously, the bodytext field, for example, was only searched
if the CType was text, textpic or textmedia. This condition is
unknown to practically all developers which has been spoken to.
The only exception so far seems to be the `EXT:mask` extension,
which already dynamically extends this condition.

It was introduced with TYPO3 v4.6 with #26829 and was pretty
much untouched ever since.

With the introduction of the TCA Schema API, we can handle the
issue much more generically for all tables with subtypes. We
can thus make the search much more precise and only search in
fields that actually exist in the respective subtype like the
CType for example.

The purpose of this search restriction is to ensure that, for
example, the content in the bodytext field is no longer found
if it was filled as a text element and then changed to a plugin.

Resolves: #104539
Related: #26829
Releases: main, 13.4
Change-Id: I490a0b25568ae1919c825553215123d73a959b44
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88094
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>