[BUGFIX] Fix PID evaluation for nested Extbase objects on creation
When a new object is persisted in extbase, the target page uid of the
object is determined either from the `persistence.storagePid` or the
`persistence.classes.<className>.newRecordStoragePid` TypoScript
configuration. This works well for top level domain objects, but not
for nested objects, which are always are added with PID 0.

This patch ensures that nested extbase objects use either the
configured storage page configuration or inherit the storage page
from the parent object if no storage page configuration is defined.

Additionally, functional tests have been added to cover the
following scenarios:

  - PID inheritance from parent objects
  - Respecting persistence storagePid configuration
  - Respecting newRecordStoragePid configuration

Resolves: #104541
Releases: main
Change-Id: Iefe8101c1bddb78e644cc6a3d40362687e4313ed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85490
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>