[BUGFIX] Fix multiple file upload handling in UploadViewHelper
When using the `f:form.upload` ViewHelper to upload files for a
property of a DTO or domain model, it is not possible to upload
multiple files using the ViewHelpers `multiple` attribute, because
the field name for form token generation is wrong. Uploading
multiple files result in a extbase property mapping exception.

The patch ensures that if the ViewHelper is configured to handle
multiple files for upload, the field name for
`registerFieldNameForFormTokenGeneration` is calculated correctly
to allow multiple files in extbase property mapping.

Resolves: #104547
Releases: main
Change-Id: I9aef3986529f9464327e8860328dd4b85b30e387
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85506
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>