[TASK] Replace column length comparison with schema normalization
The doctine/dbal integration has been raised to v4 with #102875,
but still using the Doctrine DBAL v2 way for comparing columns in
the extended `Comparator`. This kept some special treatments which
needed to be applied on the Virtual Database Scheme AND on the
actual database schema representation.

This change moves the special treatment for MySQL and MariaDB
on BlobType and TextType based columns regarding length comparison
from the TYPO3 `Comparator` into the TYPO3 `ConnectionMigrator`,
performing normalization on both schema representations now.
The normalized length does not have an impact on the raw
comparison anymore with the Doctrine DBAL 4
`AbstractPlatform::columnEqual()` check.
It is still needed to allow Doctrine determine the correct type
based on that length for the column-type SQL generation.

This change is one more step towards removing the Doctrine DBAL v3
compare-handling from the `Comparator` and further changes will
be streamlined to reach removal of legacy code.

Resolves: #104568
Related: #102875
Releases: main
Change-Id: I7094f73acb99c112ac8543c4832ec5127c7dfaf8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85552
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>