[BUGFIX] Fix type cast error in DB Compare for 'default' definition
The change in #97498 introduced a specific handling for ENUM/SET
definitions in MariaDB systems. The check may be performed with a
NULL value instead of a string, which can result in a PHP type
mismatch error.

Due to other refactoring, the issue does not exist in main (v13),
but only affects v12 and v11.

Resolves: #104621
Related: #97498
Releases: 12.4, 11.5
Change-Id: Ia28aae519b1aebcabe9a5f5c620124dfd22edc6a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85468
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>