[TASK] Decouple ext:form ConfigurationManager from extbase
Next change in the "Tame extbase ConfigurationManager" series.

ext:form comes with an add-on class of extbase
ConfigurationManager: It extends the class with an own feature
to load the relevant YAML configuration from the extensions
TypoScript. This is accompanied by an own interface.

This is not a great idea from a structural point of view:
The ext:form ConfigurationManager should be decoupled from
extbase ConfigurationManager. It may decorate it, but should
not extend it. Or - even better - not have a dependency to
extbase ConfigurationManager at all, but get the required
extbase ConfigurationManager state hand over. The latter
option is preferred since extbase ConfigurationManager
is this evil injectable stateful singleton, that hides
required state away. Turning this around in ext:form forces
consumers to hand over state directly, making the
dependency to extbase ConfigurationManager state explicit
in callers of ext:form ConfigurationManager.

The patch does this: ext:form ConfigurationManager no longer
extends extbase ConfigurationManager, but is turned into a
standalone and stateless service with an own signature that
actively receives needed state from consumers.

Resolves: #104670
Related: #104514
Releases: main
Change-Id: If6073eb50df788d45a7f4d73e95cbd2d54a8c5fe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85638
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>