[TASK] Sync ext:form FormPersistenceManagerInterface
To prepare further works on the implementing class,
FormPersistenceManagerInterface is synchronized with
some methods FormPersistenceManager implements and
are needed, but are not within the interface, yet.

Method checkForDuplicateIdentifier() is used internally
only, and removed from the interface.

Move public methods to top in FormPersistenceManager,
remove an obsolete attribute of the "fake request"
(which will see more handling with another patch), and
mark the implementing class @internal since everything
can be programmed to the interface.

Resolves: #104652
Releases: main
Change-Id: I69bf2a4118811c1407e329893559867a75457630
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85668
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>