[TASK] Update display of list views of "Backend Users" BE module
The backend module "Backend Users" has four different submodules:
'Backend users', 'Backend user groups', 'Online users' and
'Filemounts', which help the user to get an overview of the backend
users, groups and file storages available in the system.

The list views of each submodule have been adjusted to provide a more
consistent appearance and provide the user with the same user
experience as elsewhere in the system.

In the process, the locallang.xlf file has been extensively
restructured and keys renamed in order to be as well prepared as
possible for future adjustments.

Resolves: #104657
Releases: main
Change-Id: I9ec806112bd7fb5a5f81e90b07a467a78aa89c89
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85675
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>