[TASK] Reduce TYPO3_REQUEST in extbase functional tests
Various ext:extbase functional tests set
$GLOBALS['TYPO3_REQUEST']. The extbase
codebase still has various places that
access the global.

We can still reduce some usages in tests which
are obsolete meanwhile, or can be improved by
setting the request on ConfigurationManager
instead, which is one of the main callers to
the $GLOBALS['TYPO3_REQUEST'] fallback in tests.

Resolves: #104677
Related: #104675
Releases: main
Change-Id: I8672cad869e64545c482664e5d995bc1ef02bd64
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85706
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>