[BUGFIX] Make Fluid DebuggerUtility css more safe against custom css
Fix not cklickable `+` sign for the nested properties in debugger
tree in frontend.
The css selector for input fields of type checkbox are not strong enough
if the website css defines own css rules and select checkboxes with

input[type="checkbox"]

the selector in DebuggerUtility loses and properties are overriden.
Height, width, z-index and other properties were used to hide the
default appearance of checkboxes and make those form elements looking
nice. Therefore, DebuggerUtilitys selector should be a bit more
stronger.

Resolves: #104681
Releases: main, 12.4
Change-Id: Id6964e0dbf19b83318f1e5de5f3763c6be6652a0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85710
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>