[BUGFIX] Align table header in backend user/group comparison
Previously, the content of the table header in the backend user/group
comparison view was misaligned, the title was moved below the action
buttons.

This commit changes the alignment to have the title and the action
buttons next to each other. Additionally, the items are vertically
centered.

Resolves: #104691
Related: #104635
Releases: main
Change-Id: I577ee142ab6efade51810b3f72d1b2ae8d346a97
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85720
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>