[TASK] Close any open modal in InstallTool Cest tear-down
This closes any open modal in InstallTool Cest tear-down.
Thereby especially the click on 'logout' in tear-down will always work.
A modal might stay open in case the preceding test fails midway.

This makes it unnecessary to close the modal in the test-cases
itselves.

Otherwise the error reporting for the non-working click on 'logout'
might overlay the actual error.

This depends on `step_decorators: \Codeception\Step\TryTo` which is
available in `main` and `13.4`.

This implements `DataProviders` for `Application\SettingsCest`. Thereby
any flash-messages or modals from in between the test-cases are
automatically closed via `_before`/`_after`.

Resolves: #104705
Releases: main, 13.4
Change-Id: If754fc7d92725e344061d2a9c6954d3b159a8f49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87707
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>