[TASK] Do not declare extbase defaultViewObjectName, set it
Instead of declaring ActionController->defaultViewObjectName
to some value, __construct() sets it. This in line with the
recommended solution of #102632: It relaxes liskov covariance
and contravariance issues with PHP properties. Core usages now
follow to be a good example for extensions. The patch applies
minor constructor/DI cleanups along the way.

Resolves: #104770
Related: #102632
Releases: main
Change-Id: Ib013d1f52796d5e849b1e26d2173691ed7a6b310
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85822
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>