[TASK] Declare minor ext:form RendererInterface change
ext:form RendererInterface->setFormRuntime() now declares
to return void. Strictly speaking this would be a breaking
change, but it is not in practice, since extensions
typically do not implement the interface directly, but
extend AbstractElementRenderer, which follows this interface
change. Also, existing extensions could easily declare :void
(if they don't do so already), and be v13 and <v13 compatible
without problems.

Resolves: #104768
Releases: main
Change-Id: I6d012946f84212c42235346fb8abae1651a37440
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85820
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>