[TASK] Avoid FluidViewAdapter as inner view
FluidViewAdapter->setRenderingContext() sets
itself as view for inner renderings.
This was thought to be a good solution to
trigger deprecations for methods of different
view implementions (StandaloneView & friends).
However, custom fluid views are being phased
out, which will then trigger their own
deprecations.
We can thus remove this quirk.
Also, renderPartial() and renderSection() may
return mixed. This was wrong in fluid
standalone as well, and is fixed in there, too.

Resolves: #104709
Related: #101559
Releases: main
Change-Id: I381e6a26ce8b6757fc59068742cf6a09d64f8f71
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85738
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Simon Praetorius <simon@praetorius.me>