[BUGFIX] Ensure proper loading order of EXT:seo and EXT:dashboard
With #103034 EXT:styleguide received a dependency on EXT:seo.

This could lead to a situation for classic installations
(non-composer mode) where via `PackageStates.php` the
loading order was changed in a way that loaded
EXT:dashboard before EXT:seo.

This would lead to widgets defined in EXT:seo not
getting displayed in the dashboard, because the registration
would be performed too late.

The easy fix for this is to adjust the file `ext_emconf.php`
of EXT:seo to provide a 'suggest' key for EXT:dashboard.
The expected PackageStates default in the functional tests
are adapted to match that proper order.

In case EXT:seo is enabled, this 'suggest' entry ensures
that EXT:dashboard is loaded before EXT:seo, so that
everything else falls into place.

Please note that for composer installations this suggestion
was already provided.

For classic installations, [1] states:

    For non-Composer installation ("Classic Mode") the
    `ext_emconf.php` file is the source of truth for required
    dependencies and the loading order of active extensions.

NOTE: To fix a possible ordering problem in existing
installations, rebuilding `PackageStates.php` is needed.
This can be done by enabling and disabling EXT:dashboard
for example, through the Extension Manager.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/11.4/Important-95647-ComposerInstallationsAndExtensionUsage.html

Resolves: #104717
Related: #103034
Related: #95647
Releases: main, 12.4
Change-Id: I13547b0ecf6bcb60abf2e3674d273413fb6f9a6c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85750
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>