[BUGFIX] Fall back to site identifier in TypoScript overview
When TypoScript is provided by a site set, not a sys_template record and the site has no "website title" it is currently listed without any label in the "TypoScript records overview". However, the "website title" is not a mandatory field.

Fall back to the site identifier if there is not website title. The identifier is mandatory and unique.

Resolves: #104731
Releases: main
Change-Id: I7727fb966443fe55e525bdc6fa6b44a67ecbefc4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85777
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>