[BUGFIX] Allow Extbase HashService retrieval via GeneralUtility::makeInstance()
With the deprecation of extbase `HashService`, the new core
`HashService` was injected to extbase `HashService` using constructor
DI. Instantiation of extbase `HashService` however does not work,
because the service is defined by a service provider and the
required constructor argument is missing.

This change makes the extbase `HashService` injectable by DI
again by removing it from the service provider definition and
by defining it as public using `Autoconfigure` attribute.

Resolves: #104780
Related: #102763
Releases: main
Change-Id: Ife766b2306eff9ef263ba42c78dd58b1d1add27f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85827
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>