[TASK] Add AbstractMethodImplementationMatcher to extension scanner
The Extension Scanner of the EXT:install module allows to scan
custom extensions based on several static code analysis rules.

Most of these rules depend on analyzing code to see if deprecated
or removed methods are called, or if argument names change,
annotations/attributes get removed and so on.

This patch adds a `AbstractMethodImplementationMatcher` matcher which
can analyze definitions of specific methods within an extended abstract
class that need to be migrated.

A good example for this is the change in Fluid ViewHelpers deprecating
the "renderStatic" method, which could previously not be matched.

Because ViewHelpers themself do not execute a `renderStatic`
method, any custom ViewHelper implementing `renderStatic`
would not be revealed in the Extension Scanner before.

The Matcher receives a configuration like:

```
return [
 'TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper::renderStatic'
   => [
        'restFiles' => [
          'Deprecation-104789-RenderStaticForFluidViewHelpers.rst',
      ],
   ],
];
```

and based on this, iterates every custom code to check if
any class inherits from this
`TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper` abstract,
and then checks for a given method name (here `renderStatic`).
Both static (`::`) and "local" ( `->`) methods can
be configured for matching.

These matches will be reported as strong matches.

Resolves: #104799
Related: #104789
Releases: main
Change-Id: I9021ecc021a45770e2e4c8b0295f4d1afe55adfd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85853
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>