[FEATURE] Auto-add system fields showitem for tt_content
This change adds system fields - as defined in the ctrl
section - of tt_content to the showitem field list
automatically. This is either done based on palettes
that already exist or adding the fields individually
to the corresponding tabs.

Tabs that are therefore added automatically are:

* General tab at the beginning
* Language tab including language fields
* Access tab including hidden+access fields
* Description tab including descriptionColumn

In case the showitem of such type defines the
"Extended" tab, it is added at the end.

We do this for tt_content because we want a unified
way of how FormEngine renders the fields.

Resolves: #104814
Releases: main
Change-Id: I69965fbc74ae2192151ad2617c1a88cf1eb2a987
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85859
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>