[TASK] Response with MIME type application/javascript for mjs file ending
This patch adds the instructions to handle ECMAScript javascript files
'.mjs' with the correct MIME type.

The mjs file ending is already used by public packages. The different
file ending helps the distinction between modules and classic scripts.

It helps for clear communication within projects: By using the .mjs
extension, developers and build systems can easily recognize that a file
is an ECMAScript module, ensuring that the proper syntax and semantics
(such as 'import' and 'export') are used. This approach is widely common
in node projects.

While not every webserver is adding the correct MIME type for '.mjs'
it's now clearly configured within TYPO3 projects out of the box.

Resolves: #104815
Releases: main, 13.4
Change-Id: I11455ac77eb295c4505077d6453a4b3843c871a3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87615
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>