[DOCS] Add verbose renderStatic fluid ViewHelper migration steps
This is an effort to document the steps that were made in
issue #104786 to allow developers and probably integrators
to migrate affected ViewHelpers easier.

The goal is to lower unneeded fears of possibly complex
migrations, because the steps can actually be performed
mostly even with search+replace.

Resolves: #104824
Related: #104799
Related: #104789
Related: #104786
Releases: main
Change-Id: I06bd649c2ff9a4a84c31dfec9d9bcc7c6f2b81f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85893
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Stefan Bürk <stefan@buerk.tech>