[BUGFIX] Provide fallback ViewFactory for InstallTool context
The InstallTool login-notification runs in failsafe mode,
that means the symfony DI container is not booted (only
FailsafeContainer is running to avoid stale caches to
prevent interaction with the install tool).

Provide an optional fallback implementation of the
ViewFactory interface for this context, but let it be
configurable for userspace.

Resolves: #104825
Related: #104773
Releases: main
Change-Id: I566e9afff16197182ca57ab8a9cf877239373eaa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85894
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>