[TASK] Enhance Regular Expression parsing for CKEditor YAML config
The CKEditor v5 configuration API allows to specify
Regular Expression JavaScript objects, for example in
:javascript:`editor.config.typing.transformations.extra.from` or
:javascript:`editor.config.htmlSupport.allow.name`:

When TYPO3 passes YAML configuration of the CKEditor forward
to JavaScript, it uses a html-entity encoded representation,
which does not allow to utilize Regular Expression objects,
and also the CKEditor API method `buildQuotesRegExp()` is not
usable in this scenario.

This was remedied already for the configuration key :yaml:`htmlSupport`
with its sub-keys, so that when a YAML key named :yaml:`pattern`
was found, TYPO3 automatically converted that to a proper JavaScript
Regular Expression.

This is now also possible for the `editor.config.typing.transformations`
structure:

```
editor:
  config:
    typing:
      transformations:
        extra:
          - { from: { pattern: '(tsconf|t3ts)$'},
              to: 'TYPO3 TypoScript TSConfig' }
```

From searching the CKEditor v5 documentation it seems there
are no other usages where Regular Expressions are used,
but this is not easily revealed. Future places can be added
easily.

Resolves: #104827
Releases: main, 12.4
Change-Id: Ideeab2b22a177afaba2b434ef0d71920b6c7e296
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85910
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>