[TASK] Raise "phpstan/phpstan":"^1.12.3"
PHPStan released a new version containing a bug
starting to report an error if a class extends
the native PHP `\SplObjectStorage` class [1]:

    Non-abstract class ... contains abstract method
    seek() from interface SeekableIterator.

Sadly, this is a not-ignorable error and cannot be
added to the baseline.

This change upgrades the version to the faulty version
and excludes the one file meanwhile. Not good, but the
only suitable solution for now.

Used command(s):

> Build/Scripts/runTests.sh -s composer -- \
  require --dev "phpstan/phpstan":"^1.12.3"
> Build/Scripts/runTests.sh -s phpstanGenerateBaseline

[1] https://github.com/phpstan/phpstan/issues/11665

Resolves: #104866
Releases: main, 12.4
Change-Id: I8241b02310e6da3cff0292d1d37c8e69862ffb9a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85934
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>