[FEATURE] Auto-discover record preview templates
A new generic way of defining record preview template
paths is introduced. Corresponding TSconfig syntax:

    `record.preview.<table_name>.paths`

Using this option to define template paths, the
templates for the record previews as well as
related layouts and partials are auto-discovered.

Those templates receive the fully transformed record
object.

The functionality is first used for content element
previews in the page module but will be used in
other areas of the backend in the future.

Therefore, the corresponding service has been
renamed to `FluidBasedRecordPreviewRenderer`,
while the event listener method for the content
preview got the speaking name `renderPageContentPreview`.

Resolves: #104938
Releases: main
Change-Id: I15f136c50a4161366afd18f59a2188ec17750c38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86032
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>