[TASK] Hide "auto" color scheme option in dropdown
The usability of a "OS default" option isn't the greatest.
Depending on the active color scheme, the "auto" option doesn't
change anything visually. Having the imminent result options left
in the dropdown is better. The "OS default" (auto) option is still
provided in the User Settings and as default when the feature is
initially enabled (and no setting was saved by the user).

Resolves: #104942
Releases: main
Change-Id: I12e1f48446478a84069299cbb975f6fe176348e5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86051
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>