[TASK] Enrich TCA from BeforeTcaOverridesEvent
The naming of BeforeTcaOverridesEvent is a little unfortunate as it
could also be named AfterBaseTcaEvent. It was specifically meant to
enable extension authors to add base TCA in a dynamic manner. Thus,
it should also benefit from TCA enrichment.

This is important for the Content Blocks extension, as it may drop
the TCA generation of system columns and let the Core do the heavy
lifting.

Resolves: #104949
Related: #104311
Releases: main
Change-Id: I828c1b18d953071f7509ac0c9784873e03e43a8a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86078
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>